refactor: StudioCompass 에서 저장 미리보기 모달을 갈라낸다 (3,523→3,152줄)
저장소에서 가장 큰 파일의 첫 분할. 델타 미리보기 블록(미니 로즈 좌표계 ·
위성 배치 · 스크림 모달 371줄)은 나침 무대 본체와 상수도 상태도 공유하지
않는 자기완결 SVG 층이라 경계가 가장 깨끗했다. 타입만 본체에서
type-only 로 되가져온다(순환 없음). 동작 변화 0 — 공방 테스트 274 ·
계약 1,639 · e2e 8 전부 초록, tsc·lint 깨끗.
